requirements.txt
groq
python-dotenv==1.0.1
python-telegram-bot==21.5
httpx==0.28.1
aiohttp
googletrans
telegramify-markdown[mermaid]
SQLAlchemy
aiosqlite